requirements.txt
fastapi 
uvicorn 
sentence-transformers==2.2.2 
faiss-cpu 
langchain 
openai 
python-dotenv 
huggingface_hub==0.19.4
langchain<0.1.0
